Accept float fps when getting timecodes from the Timecode control otherwise inaccurate values are returned for non-integer rates. Should fix #1694.
Forward-ported from 1229de2042c785a9e851e5fc53922a75bfbf195d in master.

+++ src/wx/timecode.h
@@ -81,7 +81,7 @@ public:
 		checked_set (_fixed, t.timecode (fps));
 	}
 
-	T get (int fps) const
+	T get (float fps) const
 	{
 		T t;
 		std::string const h = wx_to_std (_hours->GetValue ());
